chore: enforce and document PR format (#4)
Adds a PR template and contribution guidelines, and ensures that the PR
title and body are formatted correctly. The title must conform to
[conventional commits](https://www.conventionalcommits.org/en/v1.0.0/)
and the body must not contain the placeholder comment, and must specify
the `Release` trailer, and if a ticket reference is present it must be
correctly formatted.

Release: false